Add phase-based camera routes with random jumps, indefinite duration, and ring wrap-copy scroll demos
- engine: widen frame_index/run_frames to u32 (indefinite runs)
- tile_scroll: surface prefetch size as template param; DESC blits for overlapping shifts; max_blit_jobs 64
- 101/102: RouteCamera with horizontal/vertical/diagonal/circular/sinusoidal phases + pauses, jump mode (2-15px) after ~1min re-randomized every 10s; 102 re-randomizes tile patterns every 10s
- 103/104: Lionheart-style ring scroll on 352x288 (2-tile margin), shifting the buffer 16px and reloading edge tiles on wrap; 103 EHB, 104 dual 3+3
- add eng/scene/route_camera.hpp

+++ engine/include/eng/engine.hpp
@@ -50,14 +50,14 @@ public:
 	///
 	/// Las demos actuales acaban tras `frame_count` para que el runner pueda capturar
 	/// y cerrar WinUAE de forma determinista. Los juegos reales tendran un bucle
-	/// controlado por estado/salida.
-	void run_frames(u16 frame_count) {
+	/// controlado por estado/salida. `0xffffffff` equivale a duracion indefinida.
+	void run_frames(u32 frame_count) {
 		GameContext context {};
 
 		m_backend.boot();
 		m_game.init(m_backend, context);
 
-		for (u16 i = 0; i < frame_count; ++i) {
+		for (u32 i = 0; i < frame_count; ++i) {
 			context.frame.frame_index = i;
 			m_game.update(m_backend, context);
 			// `render` es el punto de commit, no de simulacion. En Amiga esto importa:
